When drawing centered text, we should make sure all the text is centered relative to the frame it is being drawn in
TESTING:
Verify that if you have a fixed-width label and you want to draw centered text, that text is centered amongst itself AND it is centered relative to the label's frame.  Previously, we'd draw centered text, but that text's bounding box would be left-aligned to the label, making the text look centered (from line to line), but not actually be centered in the label.