Properly calculate animation duration when the delay or other variables have been changed
ZCAnimatedLabel used to calculate the animation duration based on the initial values of your animation delay and other settings.  If you change those values, the animation duration would not be reflected properly, so when you used the completion handler after animation, it wasn't guaranteed to trigger when you expected it to.